Prefactor: Move Project Selection List to dedicated component
